// src/app/models/market.ts
export class Market {

  position: number;
  marketName: string;
  marketCurrency: string;
  baseCurrency: string;
  marketCurrencyLong: string;
  baseCurrencyLong: string;
  minTradeSize: number;
  isActive: boolean;
  isRestricted: boolean;
  created: string;

  // "MarketCurrency": "LTC",
  // "BaseCurrency": "BTC",
  // "MarketCurrencyLong": "Litecoin",
  // "BaseCurrencyLong": "Bitcoin",
  // "MinTradeSize": 0.03297187,
  // "MarketName": "BTC-LTC",
  // "IsActive": true,
  // "IsRestricted": false,
  // "Created": "2014-02-13T00:00:00",
}